test(libsyntheticrender,libeval): split fhir-microdata, validate, profile-prompt by behaviour family
- profile-prompt.test.js (403) → -compose (composeProfilePrompt +
  hoisting) and -system-lead (composeSystemPrompt/composeLeadPrompt);
  shared RT/FIXTURES/LIVE_PROFILES lifted to -helpers.js.
- fhir-microdata.test.js (406) → -crossref (buildFhirCrossRef) and
  -render (renderFhirMicrodataHtml); shared FHIR fixture makers lifted
  to -helpers.js.
- validate.test.js (412) → -cross-content (validateCrossContent) and
  -content-validator (ContentValidator); shared buildEntities lifted to
  -helpers.js.

Assertions unchanged; test counts preserved (25, 15, 22).
Spec 0640 § C (large files), design Decision 5.
