+++ libraries/libsyntheticrender/test/fhir-microdata-render.test.js
@@ -1,201 +1,18 @@
 import { describe, test } from "node:test";
 import assert from "node:assert/strict";
+
 import {
-  renderFhirMicrodataHtml as _renderFhirMicrodataHtml,
+  renderFhirMicrodataHtml,
   buildFhirCrossRef,
-} from "../src/render/fhir-microdata.js";
-import { createDefaultRuntime } from "@forwardimpact/libutil/runtime";
-
-// Thread a real runtime (the renderer reads bundled templates via fsSync).
-const _fhirRuntime = createDefaultRuntime();
-const renderFhirMicrodataHtml = (input, config) =>
-  _renderFhirMicrodataHtml(input, config, _fhirRuntime);
-
-const PATIENT_A = "11111111-1111-4111-8111-111111111111";
-const PATIENT_B = "22222222-2222-4222-8222-222222222222";
-const PATIENT_C = "33333333-3333-4333-8333-333333333333";
-
-function makePatient(id, family, given) {
-  return {
-    resourceType: "Patient",
-    id,
-    name: [{ use: "official", family, given: [given] }],
-    gender: "female",
-    birthDate: "1980-01-01",
-  };
-}
-
-function makeCondition(patientId, code, display) {
-  return {
-    resourceType: "Condition",
-    subject: { reference: `urn:uuid:${patientId}` },
-    code: {
-      coding: [{ system: "http://snomed.info/sct", code, display }],
-      text: display,
-    },
-    onsetDateTime: "2020-01-01",
-  };
-}
-
-function makeProcedure(patientId, code, display) {
-  return {
-    resourceType: "Procedure",
-    subject: { reference: `Patient/${patientId}` },
-    code: { coding: [{ code, display }] },
-    performedDateTime: "2021-01-01",
-  };
-}
-
-function makeMedRequest(patientId, code, display) {
-  return {
-    resourceType: "MedicationRequest",
-    subject: { reference: `urn:uuid:${patientId}` },
-    medicationCodeableConcept: { coding: [{ code, display }] },
-    authoredOn: "2022-01-01",
-  };
-}
-
-function makeClinical() {
-  return {
-    conditions: [
-      {
-        id: "diabetes_t2",
-        name: "Type 2 Diabetes",
-        iri: "https://test.example/id/clinical/condition/diabetes_t2",
-      },
-    ],
-    trials: [
-      {
-        id: "oncora_p3",
-        name: "ONCORA-301",
-        conditions: ["diabetes_t2"],
-        sites: ["cambridge"],
-        iri: "https://test.example/id/clinical/trial/oncora_p3",
-      },
-    ],
-    sites: [
-      {
-        id: "cambridge",
-        iri: "https://test.example/id/clinical/site/cambridge",
-      },
-    ],
-  };
-}
-
-describe("buildFhirCrossRef", () => {
-  test("links patient to trial via condition match", () => {
-    const patients = [makePatient(PATIENT_A, "Jones", "Alice")];
-    const conditions = [
-      makeCondition(PATIENT_A, "diabetes_t2", "Type 2 Diabetes"),
-    ];
-    const crossRef = buildFhirCrossRef({
-      patients,
-      conditions,
-      clinical: makeClinical(),
-      domain: "test.example",
-    });
-
-    const patientIri = `https://test.example/id/clinical/patient/${PATIENT_A}`;
-    const trialIri = "https://test.example/id/clinical/trial/oncora_p3";
-    assert.deepStrictEqual(
-      [...(crossRef.patientToTrialIris.get(patientIri) ?? [])],
-      [trialIri],
-    );
-    assert.deepStrictEqual(
-      [...(crossRef.conditionIdToPatientIris.get("diabetes_t2") ?? [])],
-      [patientIri],
-    );
-    assert.deepStrictEqual(
-      [...(crossRef.siteIdToPatientIris.get("cambridge") ?? [])],
-      [patientIri],
-    );
-    assert.deepStrictEqual(
-      [...(crossRef.trialIriToPatientIris.get(trialIri) ?? [])],
-      [patientIri],
-    );
-  });
-
-  test("matches by display text normalization", () => {
-    const patients = [makePatient(PATIENT_A, "Jones", "Alice")];
-    const conditions = [makeCondition(PATIENT_A, "E11.9", "Type 2 Diabetes")];
-    const clinical = {
-      conditions: [{ id: "type_2_diabetes", name: "Type 2 Diabetes" }],
-      trials: [],
-      sites: [],
-    };
-    const crossRef = buildFhirCrossRef({
-      patients,
-      conditions,
-      clinical,
-      domain: "test.example",
-    });
-    const patientIri = `https://test.example/id/clinical/patient/${PATIENT_A}`;
-    assert.deepStrictEqual(
-      [...(crossRef.conditionIdToPatientIris.get("type_2_diabetes") ?? [])],
-      [patientIri],
-    );
-  });
-
-  test("returns empty maps when no conditions match", () => {
-    const patients = [makePatient(PATIENT_A, "Jones", "Alice")];
-    const conditions = [makeCondition(PATIENT_A, "unknown", "Other")];
-    const crossRef = buildFhirCrossRef({
-      patients,
-      conditions,
-      clinical: makeClinical(),
-      domain: "test.example",
-    });
-    assert.strictEqual(crossRef.patientToTrialIris.size, 0);
-    assert.strictEqual(crossRef.conditionIdToPatientIris.size, 0);
-    assert.strictEqual(crossRef.siteIdToPatientIris.size, 0);
-    assert.strictEqual(crossRef.trialIriToPatientIris.size, 0);
-  });
-
-  test("returned CrossRefIndex is frozen at the top level", () => {
-    const crossRef = buildFhirCrossRef({
-      patients: [makePatient(PATIENT_A, "Jones", "Alice")],
-      conditions: [],
-      clinical: makeClinical(),
-      domain: "test.example",
-    });
-    assert.strictEqual(Object.isFrozen(crossRef), true);
-    const before = crossRef.patientToTrialIris;
-    try {
-      crossRef.patientToTrialIris = new Map();
-    } catch {
-      /* strict mode throws; non-strict silently ignores — both produce no-op */
-    }
-    assert.strictEqual(crossRef.patientToTrialIris, before);
-  });
-
-  test("preserves insertion order across multiple patients", () => {
-    const patients = [
-      makePatient(PATIENT_A, "Jones", "Alice"),
-      makePatient(PATIENT_B, "Smith", "Bob"),
-      makePatient(PATIENT_C, "Chen", "Carol"),
-    ];
-    const conditions = [
-      makeCondition(PATIENT_A, "diabetes_t2", "Type 2 Diabetes"),
-      makeCondition(PATIENT_B, "diabetes_t2", "Type 2 Diabetes"),
-      makeCondition(PATIENT_C, "diabetes_t2", "Type 2 Diabetes"),
-    ];
-    const crossRef = buildFhirCrossRef({
-      patients,
-      conditions,
-      clinical: makeClinical(),
-      domain: "test.example",
-    });
-    const trialIri = "https://test.example/id/clinical/trial/oncora_p3";
-    assert.deepStrictEqual(
-      [...crossRef.trialIriToPatientIris.get(trialIri)],
-      [
-        `https://test.example/id/clinical/patient/${PATIENT_A}`,
-        `https://test.example/id/clinical/patient/${PATIENT_B}`,
-        `https://test.example/id/clinical/patient/${PATIENT_C}`,
-      ],
-    );
-  });
-});
+  PATIENT_A,
+  PATIENT_B,
+  PATIENT_C,
+  makePatient,
+  makeCondition,
+  makeProcedure,
+  makeMedRequest,
+  makeClinical,
+} from "./fhir-microdata-helpers.js";
 
 describe("renderFhirMicrodataHtml", () => {
   function basicInput() {
